EA-1617: JGitflow, permit override of pushReleases
Because we define this in the parent profile as "true" it becomes nearly
impossible to override this on the command line.
Instead set it to the property "pushReleases" which is what JGitFlow
uses when you don't supply it as part of the plugin configuration.

Ideally this should be a namespaced variable to avoid issues but that
requires changes to our build processes as well.

Updated changelog

Change-Id: I55ed656dda13ce7e90e2c1c856d92f72361bf995